+++ config/multi_database_commands.php
@@ -1,0 +1,11 @@
+<?php
+return [
+
+    /*
+     * Add here the list of all database connection.
+     * You can find the connection names in config file: database.connections
+     */
+    'databases' => [
+
+    ]
+];