+++ src/LaravelMultiDatabaseCommands.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace AppsInteligentes\LaravelMultiDatabaseCommands;
+
+class LaravelMultiDatabaseCommands
+{
+}